src/gemini_mcp_server/gemini_client.py
import base64
import io
import logging
from typing import Dict, Any, Optional

import google.generativeai as genai
from google.generativeai.types import HarmCategory, HarmBlockThreshold
from PIL import Image

logger = logging.getLogger(__name__)


class GeminiImageClient:
    """Client for generating images using Google Gemini API."""
    
    def __init__(self, api_key: str):
        """
        Initialize the Gemini client.
        
        Args:
            api_key: Google Gemini API key
        """
        self.api_key = api_key
        self.model = None
    
    async def initialize(self) -> None:
        """Initialize the Gemini model."""
        try:
            genai.configure(api_key=self.api_key)
            
            # Use the imagen model for image generation
            self.model = genai.GenerativeModel('gemini-1.5-flash')
            
            logger.info("Gemini client initialized successfully")
        except Exception as e:
            logger.error(f"Failed to initialize Gemini client: {e}")
            raise
    
    async def generate_image(self, prompt: str) -> Dict[str, Any]:
        """
        Generate an image from a text prompt.
        
        Args:
            prompt: Text description of the image to generate
            
        Returns:
            Dictionary containing image data and metadata
            
        Raises:
            ValueError: If the model is not initialized or prompt is invalid
            Exception: If image generation fails
        """
        if not self.model:
            raise ValueError("Gemini client not initialized")
        
        if not prompt or not prompt.strip():
            raise ValueError("Prompt cannot be empty")
        
        try:
            # For now, we'll use text generation to create a response about the image
            # Note: Gemini's image generation capabilities may vary
            # This is a placeholder implementation that would need to be updated
            # based on the actual Gemini image generation API
            
            response = await self._generate_image_placeholder(prompt)
            
            return {
                "data": response["data"],
                "mime_type": response["mime_type"],
                "prompt": prompt,
                "model": "gemini-imagen",
            }
            
        except Exception as e:
            logger.error(f"Error generating image: {e}")
            raise
    
    async def _generate_image_placeholder(self, prompt: str) -> Dict[str, Any]:
        """
        Placeholder implementation for image generation.
        
        This creates a simple colored image with the prompt text.
        In a real implementation, this would call the actual Gemini image API.
        """
        try:
            # Create a simple placeholder image
            width, height = 512, 512
            image = Image.new('RGB', (width, height), color='lightblue')
            
            # Save to bytes
            img_byte_arr = io.BytesIO()
            image.save(img_byte_arr, format='PNG')
            img_byte_arr.seek(0)
            
            # Encode to base64
            image_data = base64.b64encode(img_byte_arr.getvalue()).decode('utf-8')
            
            return {
                "data": image_data,
                "mime_type": "image/png"
            }
            
        except Exception as e:
            logger.error(f"Error creating placeholder image: {e}")
            raise
    
    def _get_safety_settings(self) -> Dict[HarmCategory, HarmBlockThreshold]:
        """Get safety settings for content generation."""
        return {
            HarmCategory.HARM_CATEGORY_HARASSMENT: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
            HarmCategory.HARM_CATEGORY_HATE_SPEECH: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
            HarmCategory.HARM_CATEGORY_SEXUALLY_EXPLICIT: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
            HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
        }
    
    async def validate_api_key(self) -> bool:
        """
        Validate the API key by making a simple request.
        
        Returns:
            True if API key is valid, False otherwise
        """
        try:
            if not self.model:
                await self.initialize()
            
            # Make a simple test request
            response = self.model.generate_content("Hello")
            return True
            
        except Exception as e:
            logger.error(f"API key validation failed: {e}")
            return False